Refactor FormGrid to vanilla extract
