+++ apps/store/src/components/PriceCalculator/PriceCalculator.tsx
@@ -13,7 +13,7 @@ import { type ShopSession } from '@/services/shopSession/ShopSession.types'
 import { AutomaticField } from './AutomaticField'
 import { FetchInsurance } from './FetchInsurance'
 import { FetchInsuranceContainer } from './FetchInsuranceContainer'
-import { FormGrid } from './FormGrid'
+import { FormGrid } from './FormGrid/FormGrid'
 import { PriceCalculatorAccordion } from './PriceCalculatorAccordion'
 import { PriceCalculatorSection } from './PriceCalculatorSection'
 import { useShowFetchInsurance } from './useFetchInsurance'
